Create Deletion_in_BST.cpp
Deleting a node in BST may be a tricky one.
There may be 3 cases: 1. Node with 0 child
                                    2. Node with 1 child
                                    3. Node with 2 child
This code includes all three possible cases.
The function takes the argument as the root node pointer and the key value that is to be deleted.
And returns the root pointer after completion.